+++ src/concurrent/mod.rs
@@ -10,9 +10,15 @@ mod lock_free_region;
 mod sharded_handle;
 mod sharded_region;
 
+#[cfg(feature = "pinning")]
+mod pinning;
+
 pub use epoch_handle::EpochHandle;
 pub use epoch_region::EpochRegion;
 pub use lock_free_handle::LockFreeHandle;
 pub use lock_free_region::LockFreeRegion;
 pub use sharded_handle::ShardedHandle;
 pub use sharded_region::ShardedRegion;
+
+#[cfg(feature = "pinning")]
+pub use pinning::PinnedRunner;